Refactored Message classes, Refactored Controllers
